+++ base/src/com/google/idea/blaze/base/bazel/BuildSystem.java
@@ -46,6 +46,8 @@ public interface BuildSystem {
   enum SyncStrategy {
     /** Never parallelize sync builds. */
     SERIAL,
+    /* Never parallelize and expand builds */
+    SERIAL_NOT_EXPAND,
     /** Parallelize sync builds if it's deemed likely that doing so will be faster. */
     DECIDE_AUTOMATICALLY,
     /** Always parallelize sync builds. */
