+++ base/src/com/google/idea/blaze/base/sync/sharding/BlazeBuildTargetSharder.java
@@ -44,10 +44,8 @@ import com.google.idea.blaze.base.sync.sharding.WildcardTargetExpander.ExpandedT
 import com.google.idea.blaze.base.sync.workspace.WorkspacePathResolver;
 import com.google.idea.common.experiments.IntExperiment;
 import com.intellij.openapi.project.Project;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
+
+import java.util.*;
 import java.util.stream.Collectors;
 
 /** Utility methods for sharding blaze build invocations. */
@@ -111,6 +109,9 @@ public class BlazeBuildTargetSharder {
 
   private static ShardingApproach getShardingApproach(
       SyncStrategy parallelStrategy, ProjectViewSet viewSet) {
+    if (parallelStrategy == SyncStrategy.SERIAL_NOT_EXPAND) {
+      return ShardingApproach.SHARD_WITHOUT_EXPANDING;
+    }
     if (shardingRequested(viewSet)) {
       return ShardingApproach.EXPAND_AND_SHARD;
     }
